+++ src/components/Header/components/HeaderLogin.test.tsx
@@ -2,36 +2,36 @@ import { render, screen, within } from "@testing-library/react";
 import { axe } from "jest-axe";
 import * as React from "react";
 
-import HeaderComponents from "./index";
-import Login from "./Login";
+import HeaderLogin from "./HeaderLogin";
+import { HeaderProvider } from "../context/headerContext";
 
-describe("Login Accessibility", () => {
+describe("HeaderLogin Accessibility", () => {
   it("passes axe accessibility test", async () => {
     const { container } = render(
-      <HeaderComponents.HeaderProvider>
-        <Login patronName="" />
-      </HeaderComponents.HeaderProvider>
+      <HeaderProvider>
+        <HeaderLogin patronName="" />
+      </HeaderProvider>
     );
     expect(await axe(container)).toHaveNoViolations();
   });
 
   it("passes axe accessibility test for mobile", async () => {
     const { container } = render(
-      <HeaderComponents.HeaderProvider>
-        <Login isMobile patronName="" />
-      </HeaderComponents.HeaderProvider>
+      <HeaderProvider>
+        <HeaderLogin isMobile patronName="" />
+      </HeaderProvider>
     );
     expect(await axe(container)).toHaveNoViolations();
   });
 });
 
-describe("Login", () => {
+describe("HeaderLogin", () => {
   describe("Desktop", () => {
     it("renders the logged out UI if there is no `patronName` value", () => {
       render(
-        <HeaderComponents.HeaderProvider>
-          <Login patronName="" />
-        </HeaderComponents.HeaderProvider>
+        <HeaderProvider>
+          <HeaderLogin patronName="" />
+        </HeaderProvider>
       );
 
       const links = screen.getAllByRole("link");
@@ -43,9 +43,9 @@ describe("Login", () => {
 
     it("renders the logged in UI if there is a `patronName` value", () => {
       render(
-        <HeaderComponents.HeaderProvider>
-          <Login patronName="PATRON, JANE A" />
-        </HeaderComponents.HeaderProvider>
+        <HeaderProvider>
+          <HeaderLogin patronName="PATRON, JANE A" />
+        </HeaderProvider>
       );
 
       const greetingContainer = screen.getByTestId("patronGreeting");
@@ -68,9 +68,9 @@ describe("Login", () => {
   describe("Mobile", () => {
     it("renders the logged out UI if there is no `patronName` value", () => {
       render(
-        <HeaderComponents.HeaderProvider>
-          <Login isMobile patronName="" />
-        </HeaderComponents.HeaderProvider>
+        <HeaderProvider>
+          <HeaderLogin isMobile patronName="" />
+        </HeaderProvider>
       );
 
       const links = screen.getAllByRole("link");
@@ -82,9 +82,9 @@ describe("Login", () => {
 
     it("renders the logged in UI if there is a `patronName` value", () => {
       render(
-        <HeaderComponents.HeaderProvider>
-          <Login isMobile patronName="PATRON, JANE A" />
-        </HeaderComponents.HeaderProvider>
+        <HeaderProvider>
+          <HeaderLogin isMobile patronName="PATRON, JANE A" />
+        </HeaderProvider>
       );
 
       const greetingContainer = screen.getByTestId("patronGreeting");