+++ src/components/Icons/IconSvgs.tsx
@@ -2,11 +2,13 @@
 import accessibilityFull from "../../../icons/svg/accessibility-full.svg";
 import accessibilityPartial from "../../../icons/svg/accessibility-partial.svg";
 import actionCheckCircle from "../../../icons/svg/action-check-circle.svg";
+import actionCheckCircleFilled from "../../../icons/svg/action-check-circle-filled.svg";
 import actionExit from "../../../icons/svg/action-exit.svg";
 import actionHelpDefault from "../../../icons/svg/action-help-default.svg";
 import actionHelpOutline from "../../../icons/svg/action-help-outline.svg";
 import actionLaunch from "../../../icons/svg/action-launch.svg";
 import actionPower from "../../../icons/svg/action-power.svg";
+import actionRegistration from "../../../icons/svg/action-registration.svg";
 import actionSettings from "../../../icons/svg/action-settings.svg";
 import alertNotificationImportant from "../../../icons/svg/alert-notification-important.svg";
 import arrow from "../../../icons/svg/arrow.svg";
@@ -47,11 +49,13 @@ export default {
   accessibilityFull,
   accessibilityPartial,
   actionCheckCircle,
+  actionCheckCircleFilled,
   actionExit,
   actionHelpDefault,
   actionHelpOutline,
   actionLaunch,
   actionPower,
+  actionRegistration,
   actionSettings,
   alertNotificationImportant,
   arrow,