+++ CHANGELOG.md
@@ -21,15 +21,25 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the mobile login menu for the `Header` component.
 - Adds the `useCloseDropDown` hook to manage closing dropdown-like components through clicks outside of it and by pressing the escape key.
 - Adds React context for the `Header`'s patron state data through `HeaderContext` and `HeaderProvider`.
-- Fixes `useNYPLBreakpoints` hook so it internally controls its own state.
+- Better namespacing for Header-only components.
 - Adds the `react-ga` package to send Google Analytics events.
 - Sets Header Auth links for production and development environments.
 
 ## Prerelease
 
+### Adds
+
+- Exports `ColorCard`.
+
 ### Updates
 
-- Updates the `Icon` component to include variants for `decorativeEnvelope`, `decorativeLibraryCard` and `decorativeShoppingBag`.
+- Updates the `Icon` component to include variants for `actionCheckCircleFilled`, `actionRegistration`, `decorativeEnvelope`, `decorativeLibraryCard` and `decorativeShoppingBag`.
+- Updates the docs for the `Colors` Style Guide to include additional data related to contrast and WCAG compliance.
+- Updates the `Checkbox` and `Radio` components to use `ui.gray.dark` for the default border color.
+
+### Fixes
+
+- Fixes `useNYPLBreakpoints` hook so it internally controls its own state. This eliminates server/client rendering issues on its initial load.
 
 ## 1.0.4 (June 23, 2022)